requirements.txt
# Core ML
torch==2.9.0
transformers==4.36.2
tokenizers==0.15.2

# Audio
librosa==0.10.2
soundfile==0.12.1

# Data — unpinned so pip picks wheels compatible with Python 3.14
numpy
pandas

# Visualisation
plotly==5.22.0

# Gemini API
google-generativeai==0.8.6

# Utilities
gdown==5.2.0
scikit-learn